requirements.txt
gspread-asyncio==1.7.1
python-telegram-bot==20.0
aiohttp==3.8.3
python-dotenv==0.20.0
pytz==2022.2.1
python-telegram-bot[job-queue]